rename name library namespace and hide unused extension, updated README.md documentation
