+++ app/src/main/java/com/devtamuno/composeblurharsh/MainActivity.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -23,11 +22,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImagePainter
-import coil.compose.rememberAsyncImagePainter
+import coil.compose.AsyncImage
 import com.devtamuno.composeblurharsh.ui.theme.ComposeBlurHarshTheme
-import com.devtamuno.composeblurhash.rememberBlurHashBrush
-import com.devtamuno.composeblurhash.rememberBlurHashPainter
+import com.devtamuno.composeblurhash.ext.rememberBlurHashPainter
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -53,8 +50,6 @@ class MainActivity : ComponentActivity() {
                         CoilImage()
                         Spacer(modifier = Modifier.height(10.dp))
                         BlurHashPainterImage()
-                        Spacer(modifier = Modifier.height(10.dp))
-                        BlurHashBrush()
 
                     }
                 }
@@ -65,8 +60,11 @@ class MainActivity : ComponentActivity() {
     @Composable
     fun CoilImage(modifier: Modifier = Modifier) {
 
-        val painter: AsyncImagePainter =
-            rememberAsyncImagePainter("https://images.unsplash.com/photo-1513002749550-c59d786b8e6c?ixlib=rb-4.0.3&q=80&fm=jpg&crop=entropy&cs=tinysrgb&w=400&fit=max")
+        val placeHolder = rememberBlurHashPainter(
+            blurString = "LvF7o6RiV@ofL4j?ozay4ptQkCfk",
+            width = 4032,
+            height = 3024,
+        )
 
         Card(
             elevation = 24.dp,
@@ -76,10 +74,12 @@ class MainActivity : ComponentActivity() {
                 .wrapContentHeight()
                 .then(modifier),
         ) {
-            Image(
-                painter = painter,
-                contentScale = ContentScale.Crop,
+            AsyncImage(
+                model = "https://images.unsplash.com/photo-1587590010936-300da0d70b9e",
                 contentDescription = null,
+                placeholder = placeHolder,
+                contentScale = ContentScale.FillBounds,
+                error = placeHolder,
                 modifier = Modifier
                     .width(250.dp)
                     .height(300.dp)
@@ -89,7 +89,6 @@ class MainActivity : ComponentActivity() {
 
     @Composable
     fun BlurHashPainterImage(modifier: Modifier = Modifier) {
-
 
         Card(
             elevation = 24.dp,
@@ -101,10 +100,9 @@ class MainActivity : ComponentActivity() {
         ) {
             Image(
                 painter = rememberBlurHashPainter(
-                    blurString = "LRG,SAxY%MRj0L%Lt6xt8^ofs:jY",
-                    width = 2848,
-                    height = 4272
-
+                    blurString = "LvF7o6RiV@ofL4j?ozay4ptQkCfk",
+                    width = 4032,
+                    height = 3024,
                 ),
                 contentScale = ContentScale.FillBounds,
                 contentDescription = null,
@@ -113,27 +111,5 @@ class MainActivity : ComponentActivity() {
                     .height(300.dp)
             )
         }
-    }
-
-    @Composable
-    fun BlurHashBrush(modifier: Modifier = Modifier) {
-
-        val rememberBlurHashBrush = rememberBlurHashBrush(
-            blurString = "LRG,SAxY%MRj0L%Lt6xt8^ofs:jY",
-            width = 2848,
-            height = 4272,
-            showBlurBackgroundImage = true
-        )
-
-        Spacer(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(300.dp)
-                .background(
-                    rememberBlurHashBrush,
-                    shape = RoundedCornerShape(10.dp)
-                )
-                .then(modifier)
-        )
     }
 }
